Флаг isFavorite теперь корректно работает (кроме промо фильма)
