+++ src/store/user/user.js
@@ -1,8 +1,6 @@
-import {addToFavorite, requireAuthorization} from '../action';
+import {requireAuthorization} from '../action';
 import {createReducer} from '@reduxjs/toolkit';
 import {AuthorizationStatus} from '../../consts/common';
-import {adaptMoviesToClient} from "../../utils/utils";
-import {getMovieList} from "../movies-data/selectors";
 
 const initialState = {
   authorizationStatus: AuthorizationStatus.NO_AUTH,
@@ -13,19 +11,6 @@ const user = createReducer(initialState, (builder) => {
     return {
       ...state,
       authorizationStatus: action.payload,
-    };
-  });
-  builder.addCase(addToFavorite, (state, action) => {
-    const movie = adaptMoviesToClient([action.payload])[0];
-    const movieIndex = getMovieList(state).findIndex((item) => item.id === movie.id);
-    const updatedMovieList = getMovieList(state).slice();
-    updatedMovieList.splice(movieIndex, 1, movie);
-
-    return {
-      ...state,
-      movie,
-      movieList: updatedMovieList,
-      isFavoriteLoaded: false,
     };
   });
 });
